Implementación del carrito: Creación de clases, repositorios y todo funcionando. Ahora falta pedido.

+++ src/main/java/com/iesvdc/demo/modelos/Camiseta.java
@@ -30,7 +30,7 @@ public class Camiseta {
     private int stock;
 
     @Column(nullable = false)
-    private double precio;
+    private float precio;
 
     @ManyToOne(optional = false)
     @JoinColumn(name = "categoria_id")
